LPDEV-830 update nugets

+++ src/Lykke.Service.PayInvoicePortal.Core/Domain/PaymentDetails.cs
@@ -61,6 +61,6 @@ namespace Lykke.Service.PayInvoicePortal.Core.Domain
 
         public DateTime? PaidDate { get; set; }
 
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
     }
 }
